Hashing

+++ Hashing.cpp
@@ -4,10 +4,10 @@ private:
 	long long base, MOD, len;
 	vector<long long>H, P;
 	string str;
-	ll mul(ll a, ll b)
+	long long mul(long long a, long long b)
 	{
 		__int128 temp = (__int128)a * b;
-		return ((ll)(temp & MOD) + (ll)(temp >> 61LL)) % MOD;
+		return ((long long)(temp & MOD) + (long long)(temp >> 61LL)) % MOD;
 	}
 	void cal()
 	{
@@ -15,7 +15,7 @@ private:
 		{
 			P[i] = mul(P[i - 1], base);
 			__int128 temp = (__int128)H[i - 1] * base + (unsigned char)str[i - 1];
-			H[i] = ((ll)(temp & MOD) + (ll)(temp >> 61LL)) % MOD;
+			H[i] = ((long long)(temp & MOD) + (long long)(temp >> 61LL)) % MOD;
 		}
 	}
 public:
@@ -24,13 +24,13 @@ public:
 		len = str.size();
 		random_device rd;
 		mt19937 engine(rd());
-		uniform_int_distribution<ll>dist((1LL << 25), (1LL << 30));
+		uniform_int_distribution<long long>dist((1LL << 25), (1LL << 30));
 		base = dist(engine);
 		H.resize(len + 2, 0);
 		P.resize(len + 2, 1);
 		cal();
 	}
-	ll get_hash(int left, int right)
+	long long get_hash(int left, int right)
 	{
 		return (H[right] - mul(H[left - 1], P[right - left + 1]) + MOD) % MOD;
 	}
